Add Monster AI in Preparation for new Spells
